requirements.txt
openai==0.28.1
tiktoken==0.6.0
streamlit
streamlit_option_menu
streamlit_extras
faiss-cpu
plotly
openpyxl
langchain
langchain_community
scipy
scikit-learn